Added support for some extra buttons / C64 keys.

+++ firmware/userio.c
@@ -52,7 +52,7 @@ void spi_uio_cmd32(unsigned char cmd, unsigned long parm) {
 }
 #endif
 
-void user_io_digital_joystick_ext(unsigned char joystick, int map) {
+void user_io_digital_joystick_ext(unsigned char joystick, uint16_t map) {
 	spi_uio_cmd8(UIO_JOYSTICK0_EXT + joystick, map);
 }
